+++ packages/i18n/messages/de.ts
@@ -241,6 +241,7 @@ Deine Daten werden niemals an weitere Parteien weitergegeben und nicht für komm
         'Deine Kontodaten, wie z.B. Profilinformationen, Erfolge und Erfahrungspunkte sowie Antworten, die du auf Fragen im KlickerUZH gibst, werden für die Lebenszeit Deines Kontos gespeichert. Deine Punkte und Platzierungen bei Kursaktivitäten und Bestenlisten werden so lange gespeichert, wie Du an der jeweiligen Kursrangliste teilnimmst. Du kannst jederzeit die Löschung Deiner Daten und Deines Kontos beantragen.',
       confirmationMessage:
         'Ich stimme den KlickerUZH [Datenschutzbestimmungen](https://www.klicker.uzh.ch/privacy_policy) und [Nutzungsbedingungen](https://www.klicker.uzh.ch/terms_of_service) zu und erkläre mich mit der darin beschriebenen Verarbeitung meiner Daten einverstanden. Mir ist bewusst, dass ich anonym und ohne Konto an den Lernaktivitäten teilnehmen kann, wenn ich diesen Bedingungen nicht zustimme.',
+      usernameAvailability: 'Dieser Benutzername ist nicht verfügbar.',
     },
     studentDocs: {
       pageList: `
@@ -356,6 +357,13 @@ Da die KlickerUZH-App noch nicht im iOS-App-Store verfügbar ist, folgen Sie die
         'Installieren Sie die KlickerUZH App auf Ihrem Handy, um Push-Benachrichtigungen zu erhalten, wenn neue Lerninhalte verfügbar sind.',
       installIOS:
         "Öffnen Sie den Share-Dialog und klicken Sie auf 'Zum Startbildschirm hinzufügen', um die KlickerUZH App auf Ihrem Handy zu installieren.",
+      createAccountJoin: 'Account erstellen & Kurs beitreten',
+      existingParticipantAccount:
+        'Sollten Sie bereits einen KlickerUZH Studierenden-Account haben, nutzen Sie bitte das Login um sich anzumelden. Einem neuen Kurs können Sie nach der Anmeldung in der Übersicht beitreten. Anderenfalls können Sie hier im Rahmen Ihres Kurses einen KlickerUZH Account erstellen.',
+      joinCourseTooltip:
+        'Geben Sie hier die PIN Ihres Kurses ein, um einen neuen Account zu erstellen und dem Kurs beizutreten. Sie erhalten diese von Ihrem Dozierenden.',
+      signup: 'Registrieren',
+      coursePinInvalid: 'Die von Ihnen eingebene Kurs-PIN ist ungültig.',
     },
     courses: {
       courseInformation: 'Kursinformationen',
@@ -927,8 +935,7 @@ Da die KlickerUZH-App noch nicht im iOS-App-Store verfügbar ist, folgen Sie die
         'Sind Sie sich sicher, dass Sie die folgende Live-Session löschen möchten?',
       liveSessionDeletionHint:
         'Das Löschen einer Live-Session ist nur möglich, solange sie nicht bereits gestartet wurde. Gelöschte Live-Sessions können nicht zu einem späteren Zeitpunkt wiederhergestellt werden.',
-      evaluationLinksEmbedding:
-        'Evaluations-Links für die Einbettung in PowerPoint',
+      evaluationLinksEmbedding: 'Links für die Einbettung der Evaluation',
       noSessions: 'Keine Sessionen gefunden',
       creationExplanation:
         'Um Ihre erste Session zu erstellen, gehen Sie zurück in den <link>Fragepool</link>. Dort können alle verschiedenen Arten von KlickerUZH-Elementen erstellt und Fragen aus dem Fragepool hinzufügen werden.',